pyproject.toml
[tool.poetry]
name = "ati-su-test-task"
version = "0.1.0"
description = ""
authors = ["Khalil <h.sultanoff@yandex.ru>"]

readme = "README.md"


[tool.poetry.dependencies]
python = ">=3.11"
asgiref = "3.8.1"
attrs = "25.3.0"
django = "5.2.1"
django-rest-framework = "0.1.0"
django-unfold = "0.59.0"
djangorestframework = "3.16.0"
drf-spectacular = "0.28.0"
inflection = "0.5.1"
jsonschema = "4.24.0"
jsonschema-specifications = "2025.4.1"
psycopg = "3.2.9"
psycopg-binary = "3.2.9"
psycopg2-binary = "2.9.10"
python-dotenv = "1.1.0"
pyyaml = "6.0.2"
referencing = "0.36.2"
rpds-py = "0.25.1"
sqlparse = "0.5.3"
typing-extensions = "4.13.2"
tzdata = "2025.2"
uritemplate = "4.1.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.2.0"
pytest = "^8.3.5"
pytest-django = "^4.11.1"


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = ["tests.py", "test_*.py"]